Update files
Now plays audio.  (Tested with ogg.  Not tested with mp3 or wav, but I think it might, maybe be good.)

---

Fixed issue with screen clearing.

---

Added a variable.

Now you can check with JS to see if the game was compiled with Quest JS.

To check:

```javascript
if(qjsPlayer){alert("This was Compiled by QuestJS");}
```

---
The audio will not loop or pause play just yet.  (I just decided to see if I could make it work at the last second.)

---
There is now Log functionality, but you have to add your own command.

Right now, it only displays an alert with the content of the log, but I will improve that.

The function is:
```javascript
showLog()
```