Update from 6.0 to 6.4.3
Updated NuGet packages that caused problems in my fork, so I deleted by fork and started from scratch.

This commit adds all of the changes from before the NuGet update.

I failed to make a changelog, anyway, so this is sort of for the greater good. (Still no changelog, but you can see all the changes in one place.)

Happy gaming!  BE SURE TO CHECK OUT THE ISSUES, BECAUSE THERE ARE MAJOR ISSUES!!!